+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 ### Deprecated
 
+* Deprecated: `open_sesame` command; use `open_sesame_open_project` instead
 * Deprecated: `open_sesame_add_folder` command; use `open_sesame_add_project` instead
 
 ## [1.5.7] - 2017-08-26